Refactoring continuation -fix docstrings -moved get_task_interval() from cumulativeresourcecalendarcpsat to schedulingcpsat -fix test and imports -cpsatauto->cpsat for shop problems, -use of WithoutNoOverlapProblem mixin for rcpsp, flex_scheduling -update some unit test to avoid checking properties not necessarly true with new cpsat solver
