+++ examples/jsp/run_cpsat.py
@@ -6,7 +6,7 @@ import logging
 
 from discrete_optimization.generic_tools.cp_tools import ParametersCp
 from discrete_optimization.shop.jsp.parser import get_data_available, parse_file
-from discrete_optimization.shop.jsp.solvers.cpsat_auto import CpSatAutoJspSolver
+from discrete_optimization.shop.jsp.solvers.cpsat import CpSatJspSolver
 
 logging.basicConfig(level=logging.INFO)
 
@@ -25,7 +25,7 @@ def run_cpsat_jsp():
         problem.n_machines,
         " machines",
     )
-    solver = CpSatAutoJspSolver(problem=problem)
+    solver = CpSatJspSolver(problem=problem)
     p = ParametersCp.default_cpsat()
     p.nb_process = 10
     res = solver.solve(parameters_cp=p, time_limit=10)
