+++ examples/shop/run_jsp.py
@@ -7,7 +7,7 @@ import logging
 from discrete_optimization.generic_tools.callbacks.loggers import ProblemEvaluateLogger
 from discrete_optimization.generic_tools.cp_tools import ParametersCp
 from discrete_optimization.shop.jsp.parser import get_data_available, parse_file
-from discrete_optimization.shop.jsp.solvers.cpsat_auto import CpSatAutoJspSolver
+from discrete_optimization.shop.jsp.solvers.cpsat import CpSatJspSolver
 from discrete_optimization.shop.solvers.greedy import GreedyShopSolver
 
 logging.basicConfig(level=logging.INFO)
@@ -27,7 +27,7 @@ def run_cpsat_jsp():
         problem.n_machines,
         " machines",
     )
-    solver = CpSatAutoJspSolver(problem=problem)
+    solver = CpSatJspSolver(problem=problem)
     p = ParametersCp.default_cpsat()
     p.nb_process = 10
     res = solver.solve(
@@ -52,7 +52,7 @@ def run_cpsat_jsp_warm_start():
     res = greedy_solver.solve()
     sol = res[-1][0]
     print("Satisfy greedy", problem.satisfy(sol))
-    solver = CpSatAutoJspSolver(problem=problem)
+    solver = CpSatJspSolver(problem=problem)
     solver.init_model()
     solver.set_warm_start(sol)
     p = ParametersCp.default_cpsat()
