+++ src/discrete_optimization/shop/fjsp/solvers/cpsat.py
@@ -1,6 +1,8 @@
 #  Copyright (c) 2026 AIRBUS and its affiliates.
 #  This source code is licensed under the MIT license found in the
 #  LICENSE file in the root directory of this source tree.
-from discrete_optimization.shop.solvers.cpsat_auto import CpSatAutoShopSolver
 
-CpSatAutoOspSolver = CpSatAutoShopSolver
+
+from discrete_optimization.shop.solvers.cpsat import CpSatShopSolver
+
+CpSatFjspSolver = CpSatShopSolver
