+++ src/discrete_optimization/shop/osp/solvers/cpsat.py
@@ -1,12 +1,6 @@
 #  Copyright (c) 2026 AIRBUS and its affiliates.
 #  This source code is licensed under the MIT license found in the
 #  LICENSE file in the root directory of this source tree.
+from discrete_optimization.shop.solvers.cpsat import CpSatShopSolver
 
-import logging
-
-from discrete_optimization.shop.solvers.cpsat_auto import CpSatAutoShopSolver
-
-logger = logging.getLogger(__name__)
-
-
-CpSatAutoJspSolver = CpSatAutoShopSolver
+CpSatOspSolver = CpSatShopSolver
