+++ tests/jsp/solvers/test_cpsat.py
@@ -13,7 +13,7 @@ from discrete_optimization.generic_tools.callbacks.early_stoppers import (
 from discrete_optimization.generic_tools.cp_tools import ParametersCp, SignEnum
 from discrete_optimization.shop.jsp.parser import get_data_available, parse_file
 from discrete_optimization.shop.jsp.problem import JobShopProblem, JobShopSolution
-from discrete_optimization.shop.jsp.solvers.cpsat_auto import CpSatAutoJspSolver
+from discrete_optimization.shop.jsp.solvers.cpsat import CpSatJspSolver
 
 
 @pytest.fixture
@@ -30,7 +30,7 @@ def problem() -> JobShopProblem:
 
 
 def test_cpsat_jsp(random_seed, problem):
-    solver = CpSatAutoJspSolver(problem=problem)
+    solver = CpSatJspSolver(problem=problem)
     parameters_cp = ParametersCp.default()
     sol: JobShopSolution = solver.solve(
         callbacks=[NbIterationStopper(nb_iteration_max=1)],
@@ -51,30 +51,24 @@ def test_cpsat_jsp(random_seed, problem):
     objective = solver.get_subtasks_makespan_variable(subtasks)
     solver.minimize_variable(objective)
     sol, _ = solver.solve(callbacks=[NbIterationStopper(nb_iteration_max=1)])[-1]
-    assert solver.solver.ObjectiveValue() == max(
-        sol.get_end_time(task) for task in subtasks
-    )
+    solver.solver.ObjectiveValue() == max(sol.get_end_time(task) for task in subtasks)
     # sum end time subtasks
     subtasks = {(0, 1), (1, 2)}
     objective = solver.get_subtasks_sum_end_time_variable(subtasks)
     solver.minimize_variable(objective)
     sol, _ = solver.solve(callbacks=[NbIterationStopper(nb_iteration_max=1)])[-1]
-    assert solver.solver.ObjectiveValue() == sum(
-        sol.get_end_time(task) for task in subtasks
-    )
+    solver.solver.ObjectiveValue() == sum(sol.get_end_time(task) for task in subtasks)
     # sum start time subtasks
     subtasks = {(0, 1), (1, 2)}
     objective = solver.get_subtasks_sum_start_time_variable(subtasks)
     solver.minimize_variable(objective)
     sol, _ = solver.solve(callbacks=[NbIterationStopper(nb_iteration_max=1)])[-1]
-    assert solver.solver.ObjectiveValue() == sum(
-        sol.get_start_time(task) for task in subtasks
-    )
+    solver.solver.ObjectiveValue() == sum(sol.get_start_time(task) for task in subtasks)
     # max end time
     objective = solver.get_global_makespan_variable()
     solver.minimize_variable(objective)
     sol, _ = solver.solve(callbacks=[NbIterationStopper(nb_iteration_max=1)])[-1]
-    assert solver.solver.ObjectiveValue() == sol.get_max_end_time()
+    solver.solver.ObjectiveValue() == sol.get_max_end_time()
 
 
 @pytest.mark.parametrize(
@@ -89,7 +83,7 @@ def test_cpsat_jsp(random_seed, problem):
     ],
 )
 def test_task_constraint(problem, task, start_or_end, sign, time, random_seed):
-    solver = CpSatAutoJspSolver(problem=problem)
+    solver = CpSatJspSolver(problem=problem)
     parameters_cp = ParametersCp.default()
     sol: JobShopSolution = solver.solve(
         parameters_cp=parameters_cp, callbacks=[NbIterationStopper(nb_iteration_max=1)]
@@ -122,7 +116,7 @@ def test_task_constraint(problem, task, start_or_end, sign, time, random_seed):
 
 
 def test_chaining_tasks_constraint(problem, random_seed):
-    solver = CpSatAutoJspSolver(problem=problem)
+    solver = CpSatJspSolver(problem=problem)
     parameters_cp = ParametersCp.default()
     task1 = (0, 0)
     task2 = (1, 4)
