+++ tests/jsp/solvers/test_dp.py
@@ -7,7 +7,7 @@ from discrete_optimization.generic_tools.callbacks.early_stoppers import (
     NbIterationStopper,
 )
 from discrete_optimization.shop.jsp.parser import get_data_available, parse_file
-from discrete_optimization.shop.jsp.solvers.cpsat_auto import CpSatAutoJspSolver
+from discrete_optimization.shop.jsp.solvers.cpsat import CpSatJspSolver
 from discrete_optimization.shop.jsp.solvers.dp import DpJspSolver
 
 
@@ -22,7 +22,7 @@ def test_dp_jsp_ws():
     # file_path = get_data_available()[1]
     file_path = [f for f in get_data_available() if "ta68" in f][0]
     problem = parse_file(file_path)
-    solver_ws = CpSatAutoJspSolver(problem)
+    solver_ws = CpSatJspSolver(problem)
     sol_ws = solver_ws.solve(time_limit=5, callbacks=[NbIterationStopper(1)])[0][0]
     solver = DpJspSolver(problem=problem)
     solver.init_model()
